+++ test/daemon/vhosts/dev.js
@@ -25,6 +25,15 @@ describe('*.tld', () => {
     })
   })
 
+  describe('GET http://any.node.dev', () => {
+    it('should proxy request', done => {
+      request(app)
+        .get('/')
+        .set('Host', 'any.node.dev')
+        .expect(200, /Hello World/, done)
+    })
+  })
+
   describe('GET http://unknown.dev', () => {
     it('should return 404', done => {
       request(app)